[build] Remove linkage of logging_filter -> otel
