+++ BUILD
@@ -2372,6 +2372,7 @@ grpc_cc_library(
         "//src/core:channel_stack_type",
         "//src/core:context",
         "//src/core:error",
+        "//src/core:logging_filter",
         "//src/core:slice",
         "//src/core:slice_buffer",
         "//src/core:slice_refcount",
