+++ src/core/ext/filters/logging/logging_filter.h
@@ -22,8 +22,38 @@
 #include <grpc/support/port_platform.h>
 
 #include "src/core/ext/filters/logging/logging_sink.h"
+#include "src/core/lib/channel/promise_based_filter.h"
 
 namespace grpc_core {
+
+class ClientLoggingFilter final : public ChannelFilter {
+ public:
+  static const grpc_channel_filter kFilter;
+
+  static absl::StatusOr<ClientLoggingFilter> Create(
+      const ChannelArgs& args, ChannelFilter::Args /*filter_args*/);
+
+  // Construct a promise for one call.
+  ArenaPromise<ServerMetadataHandle> MakeCallPromise(
+      CallArgs call_args, NextPromiseFactory next_promise_factory) override;
+
+ private:
+  explicit ClientLoggingFilter(std::string default_authority)
+      : default_authority_(std::move(default_authority)) {}
+  std::string default_authority_;
+};
+
+class ServerLoggingFilter final : public ChannelFilter {
+ public:
+  static const grpc_channel_filter kFilter;
+
+  static absl::StatusOr<ServerLoggingFilter> Create(
+      const ChannelArgs& args, ChannelFilter::Args /*filter_args*/);
+
+  // Construct a promise for one call.
+  ArenaPromise<ServerMetadataHandle> MakeCallPromise(
+      CallArgs call_args, NextPromiseFactory next_promise_factory) override;
+};
 
 void RegisterLoggingFilter(LoggingSink* sink);
